refactor b09 to multiple files
