+++ coco/b09/procbank.py
@@ -2,7 +2,7 @@ import importlib.resources as pkg_resources
 import re
 from collections import defaultdict
 
-from . import resources
+from .. import resources
 
 # Procedure names that start with a procedure keyword.
 PROCEDURE_START_PREFIX = re.compile(r'(?i)procedure\s+(\w+)\s*$')
