+++ examples/decb/width.bas
@@ -1,0 +1,4 @@
+10 WIDTH 80
+20 LOCATE 32,10
+30 PRINT "hello world"
+40 PRINT @65,"wow, this works"
